Adapt GerritIndexer to be used with ElasticSearch
There was no GerritIndexer implementation supporting Elasticsearch.
However, even with a shared index, it makes sense to have a separate
process/pod to perform a change index schema upgrade to avoid
performance issues for clients.

If elasticsearch is configured, the GerritIndexer will directly
create the new index in Elasticsearch and not attempt to copy
indexes.

The index configuration in the GerritCluster can be overridden
for the GerritIndexer. This is useful in migration scenarios, where
one would want to continue to use the previous index type, while
preparing the new one.

+++ container-images/gerrit-init/tools/gerrit-initializer/initializer/tasks/reindex.py
@@ -200,6 +200,9 @@ class GerritElasticSearchReindexer(GerritAbstractReindexer):
     def _prepare(self):
         index_mnt_path = f"{MNT_PATH}/index"
         index_site_path = f"{SITE_PATH}/index"
+        if not os.path.exists(index_mnt_path):
+            LOG.warning("No mounted index directory found in %s", index_mnt_path)
+            return
         if os.path.exists(index_site_path):
             if os.path.islink(index_site_path):
                 os.unlink(index_site_path)
